+++ packaging/chocolatey/tools/chocolateyinstall.ps1
@@ -4,7 +4,7 @@
 $ErrorActionPreference = 'Stop'
 
 $packageName = 'empirica'
-$packageVersion = '1.12.30'
+$packageVersion = '1.12.31'
 $url = "https://files.pythonhosted.org/packages/source/e/empirica/empirica-$packageVersion.tar.gz"
 $checksum = 'd05c212e17e7bbaf6ac48a53f68d229d5e79bd6eb9580bce12727fb2fe31657f'  # TODO: Update sha256 after PyPI publish
 $checksumType = 'sha256'